Move useCsbData() and useBapSamData() custom hooks to utilities file
