+++ app/client/src/routes/newApplicationForm.tsx
@@ -5,17 +5,17 @@ import { XMarkIcon } from "@heroicons/react/24/outline";
 import icons from "uswds/img/sprite.svg";
 // ---
 import { serverUrl, messages, postData } from "../config";
-import { getUserInfo } from "../utilities";
+import {
+  BapSamEntity,
+  useCsbData,
+  useBapSamData,
+  getUserInfo,
+} from "../utilities";
 import { Loading } from "components/loading";
 import { Message } from "components/message";
 import { MarkdownContent } from "components/markdownContent";
 import { TextWithTooltip } from "components/tooltip";
 import { useContentData } from "components/app";
-import {
-  BapSamEntity,
-  useCsbData,
-  useBapSamData,
-} from "components/userDashboard";
 
 type FormioSubmission = {
   [field: string]: unknown;
