+++ app/client/src/utilities.tsx
@@ -1,4 +1,60 @@
-import { BapSamEntity } from "components/userDashboard";
+import { useQueryClient } from "@tanstack/react-query";
+
+export type CsbData = {
+  submissionPeriodOpen: {
+    application: boolean;
+    paymentRequest: boolean;
+    closeOut: boolean;
+  };
+};
+
+export type BapSamEntity = {
+  ENTITY_COMBO_KEY__c: string;
+  UNIQUE_ENTITY_ID__c: string;
+  ENTITY_EFT_INDICATOR__c: string;
+  ENTITY_STATUS__c: "Active" | string;
+  LEGAL_BUSINESS_NAME__c: string;
+  PHYSICAL_ADDRESS_LINE_1__c: string;
+  PHYSICAL_ADDRESS_LINE_2__c: string | null;
+  PHYSICAL_ADDRESS_CITY__c: string;
+  PHYSICAL_ADDRESS_PROVINCE_OR_STATE__c: string;
+  PHYSICAL_ADDRESS_ZIPPOSTAL_CODE__c: string;
+  PHYSICAL_ADDRESS_ZIP_CODE_4__c: string;
+  // contacts
+  ELEC_BUS_POC_EMAIL__c: string | null;
+  ELEC_BUS_POC_NAME__c: string | null;
+  ELEC_BUS_POC_TITLE__c: string | null;
+  //
+  ALT_ELEC_BUS_POC_EMAIL__c: string | null;
+  ALT_ELEC_BUS_POC_NAME__c: string | null;
+  ALT_ELEC_BUS_POC_TITLE__c: string | null;
+  //
+  GOVT_BUS_POC_EMAIL__c: string | null;
+  GOVT_BUS_POC_NAME__c: string | null;
+  GOVT_BUS_POC_TITLE__c: string | null;
+  //
+  ALT_GOVT_BUS_POC_EMAIL__c: string | null;
+  ALT_GOVT_BUS_POC_NAME__c: string | null;
+  ALT_GOVT_BUS_POC_TITLE__c: string | null;
+  //
+  attributes: { type: string; url: string };
+};
+
+export type BapSamData =
+  | { results: false; entities: [] }
+  | { results: true; entities: BapSamEntity[] };
+
+/** Custom hook that returns cached fetched CSB data */
+export function useCsbData() {
+  const queryClient = useQueryClient();
+  return queryClient.getQueryData<CsbData>(["csb-data"]);
+}
+
+/** Custom hook that returns cached fetched BAP SAM.gov data */
+export function useBapSamData() {
+  const queryClient = useQueryClient();
+  return queryClient.getQueryData<BapSamData>(["bap-sam-data"]);
+}
 
 /**
  * Returns a user’s title and name when provided an email address and a SAM.gov
